pyproject.toml
[project]
name = "bpjs-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "mysql-connector-python>=9.4.0",
    "ollama>=0.5.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
]
